requirements.txt
Pillow~=9.4.0
cryptography~=39.0.0